# API_Document/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "如意外卖接口文档",
    "version": "1.0",
    "title": "如意外卖"
  },
  "host": "localhost:9000",
  "basePath": "/",
...
  "paths": {
    "/addressBook": {
      "post": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "添加地址信息",
        "operationId": "saveUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook对象»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "修改地址信息",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "删除地址",
        "operationId": "deleteUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/addressBook/default": {
      "get": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "获取默认地址",
        "operationId": "getDefaultUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook对象»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "设置默认地址",
        "operationId": "setDefaultUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook对象»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/addressBook/list": {
      "get": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "获取指定地址",
        "operationId": "listUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cityCode",
            "in": "query",
            "description": "市级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "cityName",
            "in": "query",
            "description": "市级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "consignee",
            "in": "query",
            "description": "收货人",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "detail",
            "in": "query",
            "description": "详细地址",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtCode",
            "in": "query",
            "description": "区级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtName",
            "in": "query",
            "description": "区级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "isDefault",
            "in": "query",
            "description": "默认 0 否 1是",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "label",
            "in": "query",
            "description": "标签",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceCode",
            "in": "query",
            "description": "省级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceName",
            "in": "query",
            "description": "省级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "sex",
            "in": "query",
            "description": "性别 0 女 1 男",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "用户id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«AddressBook对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/addressBook/{id}": {
      "get": {
        "tags": [
          "地址铺管理"
        ],
        "summary": "获取地址列表",
        "operationId": "getUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/category": {
      "post": {
        "tags": [
          "分类管理"
        ],
        "summary": "新增分类",
        "operationId": "saveUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "description": "category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "分类管理"
        ],
        "summary": "修改分类",
        "operationId": "updateUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "description": "category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "分类管理"
        ],
        "summary": "删除分类",
        "operationId": "deleteUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/category/list": {
      "get": {
        "tags": [
          "分类管理"
        ],
        "summary": "查询分类数据",
        "operationId": "listUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "description": "分类名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "顺序",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "type",
            "in": "query",
            "description": "类型   1 菜品分类 2 套餐分类",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Category对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/category/page": {
      "get": {
        "tags": [
          "分类管理"
        ],
        "summary": "分页查询套餐及菜品分类",
        "operationId": "pageUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddressBook对象": {
      "type": "object",
      "properties": {
        "cityCode": {
          "type": "string",
          "description": "市级区划编号",
          "refType": null
        },
        "cityName": {
          "type": "string",
          "description": "市级名称",
          "refType": null
        },
        "consignee": {
          "type": "string",
          "description": "收货人",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "detail": {
          "type": "string",
          "description": "详细地址",
          "refType": null
        },
        "districtCode": {
          "type": "string",
          "description": "区级区划编号",
          "refType": null
        },
        "districtName": {
          "type": "string",
          "description": "区级名称",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "isDefault": {
          "type": "integer",
          "format": "int32",
          "description": "默认 0 否 1是",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "label": {
          "type": "string",
          "description": "标签",
          "refType": null
        },
        "phone": {
          "type": "string",
          "description": "手机号",
          "refType": null
        },
        "provinceCode": {
          "type": "string",
          "description": "省级区划编号",
          "refType": null
        },
        "provinceName": {
          "type": "string",
          "description": "省级名称",
          "refType": null
        },
        "sex": {
          "type": "string",
          "description": "性别 0 女 1 男",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "用户id",
          "refType": null
        }
      },
      "title": "AddressBook对象",
      "description": "地址管理"
    },
    "Category对象": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "分类名称",
          "refType": null
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "顺序",
          "refType": null
        },
        "type": {
          "type": "integer",
          "format": "int32",
          "description": "类型   1 菜品分类 2 套餐分类",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "Category对象",
      "description": "菜品及套餐分类"
    },
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
...
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
    "R«AddressBook对象»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/AddressBook对象"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«AddressBook对象»"
    },
...
    "R«List«AddressBook对象»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddressBook对象"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«AddressBook对象»»"
    },
    "R«List«Category对象»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category对象"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«Category对象»»"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
...
    "R«object»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "object"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«object»"
    },
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
...
  }
}